+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/account/SocialLoginAPI.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.account.auth.controller;
+package com.mobble.mobbleserver.infrastructure.web.account;
 
 import com.mobble.mobbleserver.account.auth.dto.request.SocialLoginRequestDto;
 import com.mobble.mobbleserver.account.auth.dto.response.SocialLoginResponseDto;
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/auth")
-public class SocialLoginController {
+public class SocialLoginAPI {
 
     private final SocialLoginService socialLoginService;
